requirements.txt
keyboard==0.13.5
PyAutoGUI==0.9.53
PySide2==5.15.2
python-mpv==0.5.2
